+++ comps/cores/mega/http_service.py
@@ -8,6 +8,8 @@ from typing import Optional
 from fastapi import FastAPI
 from prometheus_fastapi_instrumentator import Instrumentator
 from uvicorn import Config, Server
+import asyncio
+import multiprocessing
 
 from .base_service import BaseService
 from .base_statistics import collect_all_statistics
@@ -83,6 +85,11 @@ class HTTPService(BaseService):
 
         return app
 
+    def add_startup_event(self, func):
+        @self.app.on_event("startup")
+        async def startup_event():
+            asyncio.create_task(func)
+
     async def initialize_server(self):
         """Initialize and return HTTP server."""
         self.logger.info("Setting up HTTP server")
@@ -110,11 +117,9 @@ class HTTPService(BaseService):
                 """
                 await self.main_loop()
 
-        app = self.app
-
         self.server = UviServer(
             config=Config(
-                app=app,
+                app=self.app,
                 host=self.host_address,
                 port=self.primary_port,
                 log_level="info",
@@ -136,6 +141,35 @@ class HTTPService(BaseService):
         self.server.should_exit = True
         await self.server.shutdown()
         self.logger.info("Server termination completed")
+
+    def _async_setup(self):
+        self.event_loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(self.event_loop)
+        self.event_loop.run_until_complete(self.initialize_server())
+
+    def run(self):
+        """Running method to block the main thread.
+
+        This method runs the event loop until a Future is done. It is designed to be called in the main thread to keep it busy.
+        """
+        self.event_loop.run_until_complete(self.execute_server())
+
+    def start(self, in_single_process=False):
+        print(in_single_process)
+        if in_single_process:
+            # Resolve HPU segmentation fault and potential tokenizer issues by limiting to same process
+            self.run()
+        else:
+            self.process = multiprocessing.Process(target=self.run, daemon=False, name=self.name)
+            self.process.start()
+
+    def stop(self):
+        self.event_loop.run_until_complete(self.terminate_server())
+        self.event_loop.stop()
+        self.event_loop.close()
+        self.server.logger.close()
+        if self.process.is_alive():
+            self.process.terminate()
 
     @staticmethod
     def check_server_readiness(ctrl_address: str, timeout: float = 1.0, logger=None, **kwargs) -> bool:
@@ -170,3 +204,6 @@ class HTTPService(BaseService):
         :return: True if status is ready else False.
         """
         return HTTPService.check_server_readiness(ctrl_address, timeout, logger=logger)
+
+    def add_route(self, endpoint, handler, methods=["POST"]):
+        self.app.router.add_api_route(endpoint, handler, methods=methods)
